Added text fade on main menu
Also fixed a few bugs

+++ Assets/Scripts/MainMenu.cs
@@ -9,7 +9,7 @@ public class MainMenu : MonoBehaviour
 {
     [SerializeField] private TMP_Text energyText;
     [SerializeField] private int maxEnergy;
-    [SerializeField] private int energyRechargeDuration;
+    [SerializeField] private float energyRechargeDuration;
     [SerializeField] private AndroidNotificationHandler androidNotificationHandler;
